Reform202504

+++ test/tq_11_realign_bypass.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-### THAME-Q tq_10_realign.sh
+### THAME-Q tq_11_realign_bypass.sh ID
 ### Objectives:
 # This script is designed to realign MRI (T1-weighted) and dynamic PET images to a standard brain template (MNI152).
 # The process includes brain extraction, rigid body transformation, and realignment of individual PET frames.
@@ -36,10 +36,14 @@ source ${THAMEQDIR}/config.env
 
 # Create a basis for coregistration QC
 QCT1W=./coregistration_results_t1w.csv
-echo "ID,Rx,Ry,Rz,Dice" > ${QCT1W}
+if [[ ! -e ${QCT1W} ]]; then
+  echo "ID,Rx,Ry,Rz,Dice" > ${QCT1W}
+fi
 
 QCPET=./coregistration_results_pet.csv
-echo "ID,Rmax_frame,Rx_mean,Ry_mean,Rz_mean,Dice" > ${QCPET}
+if [[ ! -e ${QCPET} ]]; then
+  echo "ID,Rmax_frame,Rx_mean,Ry_mean,Rz_mean,Dice" > ${QCPET}
+fi
 
 # Define util function
 function calc_dice() {
@@ -48,14 +52,15 @@ function calc_dice() {
   echo "scale=3;$overlap*2/$union" | bc
 }
 
-for f in [A-Z]*_pmpbb3_dyn.nii*
+for id in "$@"
 do
+  id=${id%_t1w.nii*}
+  f=${id%_pmpbb3_dyn.nii*}_pmpbb3_dyn.nii.gz
   pet=$(imglob $f) # PET filename
   t1w=${pet/pmpbb3_dyn/t1w} # T1 filename
   ref=${FSLDIR}/data/standard/MNI152_T1_1mm_brain # Template path
   pad=MNI152_T1_1mm_pad # padded images
-
-  echo "Process: ${t1w%_t1w}"
+  petref=${pet}_ref # reference image for PET realignment
   
   ### Create MNI152_T1_1mm_pad with the same FOV of the T1 image.
   # Calculate the required number of voxels for each side.
@@ -72,68 +77,66 @@ do
   fslroi ${ref} ${pad} ${xmin} ${xsize} ${ymin} ${ysize} ${zmin} ${zsize}
   
   # Reorient T1
-  echo "Reorient ${t1w}"
-  fslreorient2std ${t1w} ${t1w}_o
-  
-  # Rigid body transform of T1 to MNI
-  echo "Rigid body transform of ${t1w} to MNI"
-  #bet ${t1w}_o ${t1w}_brain -R -B -f 0.20 # Brain Extraction
-  mri_synthstrip -i ${t1w}_o.nii -o ${t1w}_brain.nii -m ${t1w}_brain_mask.nii
-  flirt -dof 6 -in ${t1w}_brain -ref ${pad} -omat ${t1w}2MNI.mat -out ${t1w}_brain_r # Conversion matrix from native space to MNI
-  flirt -dof 6 -in ${t1w}_o -ref ${pad} -applyxfm -init ${t1w}2MNI.mat -out ${t1w}_r # Realign T1 image to MNI template
-  flirt -dof 6 -in ${t1w}_brain_mask.nii -ref ${pad} -interp nearestneighbour -applyxfm -init ${t1w}2MNI.mat -out ${t1w}_brain_mask_r.nii
-  
-  # Evaluation for T1 x MNI coregistratioin
-  mri_synthstrip -i ${pad}.nii -m ${t1w%_t1w}_mnipad_stripmask.nii.gz
-  DICE_T1W=$(calc_dice ${t1w}_brain_mask_r.nii ${t1w%_t1w}_mnipad_stripmask.nii.gz)
-  R_T1W=$(avscale --allparams ${t1w}2MNI.mat | grep 'Rotation Angles' | awk -F '= ' '{print $2}' | sed 's/ /,/g')
-  echo "${t1w%_t1w},${R_T1W%,},$DICE_T1W" >> ${QCT1W}
-
-  ## Split PET frames as ${pet}_f????.nii
-  echo "Split PET frames"
-  fslsplit ${pet} ${pet}_f
-  
-  # Calculate a mean image from first two images of PET
-  #echo "Calculate a mean image of PET as target"
-  #if [[ $(ls | grep ${pet}_f) >2 ]]; then
-  #  fslroi ${pet} ${pet}_two 0 2 
-  #  fslmaths ${pet}_two -Tmean ${pet}_ref  
-  #  petref=${pet}_ref
-  #else
-  #  petref=${pet}_f0000
-  #fi
-  petref=${pet}_f0000
-
-  ## PET frames are realigned, averaged, and coregistered to T1W
-  echo "Realign each PET frame to target image\nTarget: ${petref}"
-  
-  # Set MAXRUNNING
-  CPU_LIMIT=$(( $(nproc) - 1 ))
-  if [[ "$CPU_LIMIT" -lt 1 ]]; then CPU_LIMIT=1; fi
-
-  TOTAL_MEM_MB=$(free -m | awk '/^Mem:/{print $2}')
-  RAM_LIMIT=$(( TOTAL_MEM_MB / 2048 ))
-  if [ "$RAM_LIMIT" -lt 1 ]; then RAM_LIMIT=1; fi
-
-  if [[ "$CPU_LIMIT" -le "$RAM_LIMIT" ]]; then
-    MAX_JOBS=$CPU_LIMIT
-  else
-    MAX_JOBS=$RAM_LIMIT
+  if [[ ! -e ${t1w}_r.nii ]]; then
+    echo "Reorient ${t1w}"
+    fslreorient2std ${t1w} ${t1w}_o
+    
+    # Rigid body transform of T1 to MNI
+    echo "Rigid body transform of ${t1w} to MNI"
+    #bet ${t1w}_o ${t1w}_brain -R -B -f 0.20 # Brain Extraction
+    mri_synthstrip -i ${t1w}_o.nii -o ${t1w}_brain.nii -m ${t1w}_brain_mask.nii
+    flirt -dof 6 -in ${t1w}_brain -ref ${pad} -omat ${t1w}2MNI.mat -out ${t1w}_brain_r # Conversion matrix from native space to MNI
+    flirt -dof 6 -in ${t1w}_o -ref ${pad} -applyxfm -init ${t1w}2MNI.mat -out ${t1w}_r # Realign T1 image to MNI template
+    flirt -dof 6 -in ${t1w}_brain_mask.nii -ref ${pad} -interp nearestneighbour -applyxfm -init ${t1w}2MNI.mat -out ${t1w}_brain_mask_r.nii
+    
+    # Evaluation for T1 x MNI coregistratioin
+    mri_synthstrip -i ${pad}.nii -m ${t1w%_t1w}_mnipad_stripmask.nii.gz
+    DICE_T1W=$(calc_dice ${t1w}_brain_mask_r.nii ${t1w%_t1w}_mnipad_stripmask.nii.gz)
+    R_T1W=$(avscale --allparams ${t1w}2MNI.mat | grep 'Rotation Angles' | awk -F '= ' '{print $2}' | sed 's/ /,/g')
+    echo "${t1w%_t1w},${R_T1W%,},$DICE_T1W" >> ${QCT1W}
   fi
 
-  for t in ${pet}_f*[0-9].nii
-  do
-    if [[ "$t" = "${pet}_f0000.nii" ]]; then
-      cp ${pet}_f0000.nii ${pet}_f0000_align.nii
-      continue
+  if [[ ! -e ${pet}_align_mean_fmean.nii ]]; then    
+    ## Split PET frames as ${pet}_f????.nii
+    echo "Split PET frames"
+    fslsplit ${pet} ${pet}_f
+    
+    # Calculate a mean image from first two images of PET
+    echo "Calculate a mean image of PET as target"
+    if [[ $(ls | grep ${pet}_f) >2 ]]; then
+      fslroi ${pet} ${pet}_two 0 2 
+      fslmaths ${pet}_two -Tmean ${petref}  
+    else
+      cp ${pet}_f0000.nii ${petref}
     fi
 
-    while [[ "$(jobs -rp | wc -l)" -ge "$MAX_JOBS" ]]; do
-      sleep 10s
-    done
+    ## PET frames are realigned, averaged, and coregistered to T1W
+    echo "Realign each PET frame to target image"
     
-    flirt -dof 6 -in $t -ref ${petref} -cost normmi -searchcost normmi -omat ${t%.nii}_align.mat -out ${t%.nii}_align &  
-  done
+    # Set MAXRUNNING
+    CPU_LIMIT=$(( $(nproc) - 1 ))
+    if [[ "$CPU_LIMIT" -lt 1 ]]; then CPU_LIMIT=1; fi
+
+    TOTAL_MEM_MB=$(free -m | awk '/^Mem:/{print $2}')
+    RAM_LIMIT=$(( TOTAL_MEM_MB / 2048 ))
+    if [ "$RAM_LIMIT" -lt 1 ]; then RAM_LIMIT=1; fi
+
+    if [[ "$CPU_LIMIT" -le "$RAM_LIMIT" ]]; then
+      MAX_JOBS=$CPU_LIMIT
+    else
+      MAX_JOBS=$RAM_LIMIT
+    fi
+
+    for t in ${pet}_f*[0-9].nii
+    do
+      while [[ "$(jobs -rp | wc -l)" -ge "$MAX_JOBS" ]]; do
+        sleep 10s
+      done
+      
+      flirt -dof 6 -in $t -ref ${petref} -cost normmi -searchcost normmi -omat ${t%.nii}_align.mat -out ${t%.nii}_align &  
+    done
+
+  fi
 
   wait
   
@@ -143,17 +146,19 @@ do
   Rmaxf=$(for v in $Rf; do echo $v; done | sort -nr | head -n1)
 
   # Merge realigned frames and mean them
-  echo "Merge realigned frames"
-  fslmerge -t ${pet}_align ${pet}_f*_align.nii
-  fslmaths ${pet}_align -Tmean ${pet}_align_mean
-  #fslmaths ${pet}_align_mean -fmean ${pet}_align_mean_fmean
+  if [[ ! -e ${pet}_align_mean_fmean.nii ]]; then
+    echo "Merge realigned frames"
+    fslmerge -t ${pet}_align ${pet}_f*_align.nii
+    fslmaths ${pet}_align -Tmean ${pet}_align_mean
+    fslmaths ${pet}_align_mean -fmean ${pet}_align_mean_fmean
+  fi
+  
+  # Coregistration with Brain Extraction
   #flirt -dof 6 -in ${pet}_align_mean -ref ${t1w}_r -cost normmi -searchcost normmi -omat ${t1w%_t1w}_PET2T1W.mat -out ${pet}_mean
-  fslmaths ${pet}_align_mean -thr 0 -bin -fillh -ero ${pet}_align_headmask
-  fslmaths ${pet}_align_mean -mas ${pet}_align_headmask ${pet}_align_mean_head
-
-  echo "Coregister the realigned and averaged PET to T1w"
-  flirt -dof 6 -in ${pet}_align_mean_head -ref ${t1w}_r -cost normmi -searchcost normmi -omat ${t1w%_t1w}_PET2MNI.mat
-  flirt -dof 6 -in ${pet}_align_mean -ref ${t1w}_r -applyxfm -init ${t1w%_t1w}_PET2MNI.mat -out ${pet}_mean
+  mri_synthstrip -i ${pet}_align_mean_fmean.nii -o ${pet}_align_mean_strip.nii.gz
+  flirt -dof 6 -in ${pet}_align_mean_strip -ref ${t1w}_brain_r -cost normmi -searchcost normmi -omat ${t1w%_t1w}_PET2MNI.mat
+  #flirt -dof 6 -in ${pet}_align_mean_stripmask -ref ${t1w}_brain_r -interp nearestneighbour -applyxfm -init ${t1w%_t1w}_PET2T1W.mat -out ${pet}_mean_stripmask
+  flirt -dof 6 -in ${pet}_align_mean -ref ${t1w}_brain_r -applyxfm -init ${t1w%_t1w}_PET2MNI.mat -out ${pet}_mean
 
   # Calculate mean of realigned PET images and
   # divide the image by voxel size to produce kbq/cc image
@@ -166,7 +171,7 @@ do
   mri_synthstrip -i ${pet}_mean.nii -m ${pet}_mean_stripmask.nii.gz
   DICE_PET=$(calc_dice ${t1w}_brain_mask_r.nii ${pet}_mean_stripmask.nii.gz)
   R_PET=$(avscale --allparams ${t1w%_t1w}_PET2MNI.mat | grep 'Rotation Angles' | awk -F '= ' '{print $2}' | sed 's/ /,/g')
-  echo "${t1w%_t1w},$Rmaxf,${R_PET%,},$DICE_PET" >> ${QCPET}
+  echo "${t1w%_t1w},$Rmaxf,${R_PET%,},$DICE_PET,BYPASS" >> ${QCPET}
 
   # Create QA Report
   fslmaths ${t1w}_brain_mask_r.nii -ero tmpmask
@@ -174,7 +179,7 @@ do
   rm tmpmask.nii
 
   convert_xfm -omat ${t1w%_t1w}_MNI2PET.mat -inverse ${t1w%_t1w}_PET2MNI.mat
-  flirt -dof 6 -in ${t1w}_brain_outline_r -ref ${pet}_align_mean_head -interp nearestneighbour -applyxfm -init ${t1w%_t1w}_MNI2PET.mat -out ${t1w}_brain_outline4pet
+  flirt -dof 6 -in ${t1w}_brain_outline_r -ref ${pet}_align_mean_strip -interp nearestneighbour -applyxfm -init ${t1w%_t1w}_MNI2PET.mat -out ${t1w}_brain_outline4pet
 
   #fslmaths ${pet}_mean_stripmask -ero tmpmask
   #fslmaths ${pet}_mean_stripmask -sub tmpmask ${pet}_mean_outline
